requirements.txt
torch
torchvision
diffusers
transformers
accelerate
opencv-python-headless
pillow
controlnet_aux
fastapi
uvicorn[standard]
python-multipart